move subscription components into a subfolder
