+++ components/settings/subscription/components/BlocksExplanation.tsx
